♻️ refactor [#10.9.3]: 2차 개선 - SSR 안전성 강화
🔧 변경 사항:
- **[Modular]**: 로거, 타입, 설정 모듈 분리
- **[SSR Safety]**: WebSocket 상수를 숫자로 변경
- **[Configurable]**: SSR 폴백 URL 환경 변수화
- **[Simplified]**: URL 검증 로직 단순화

📁 새로운 파일:
- web_ui/src/utils/logger.ts (재사용 가능한 로거)
- web_ui/src/types/websocket.ts (WebSocket 타입 및 상수)

♻️ 리팩토링:
- web_ui/src/config/websocket.ts (설정에만 집중)

🛡️ SSR 안전성:
1. **WebSocket 상수 문제 해결**:
   - `WebSocket.CONNECTING` → `WS_READY_STATE.CONNECTING` (0)
   - SSR 환경에서 `WebSocket is undefined` 오류 방지

2. **SSR 폴백 URL 설정 가능**:
   - `NEXT_PUBLIC_WS_FALLBACK_URL` 환경 변수 추가
   - 프로덕션 SSR에서 localhost 하드코딩 제거

3. **URL 검증 단순화**:
   - `new URL()` try/catch → 정규식 검증
   - 가독성 및 성능 향상

4. **JSDoc 정확성**:
   - `@throws` 제거 (실제 동작과 불일치)
   - URL 우선순위 명시

📚 모듈 구조:
- `utils/logger.ts`: 중앙 집중식 로깅
- `types/websocket.ts`: WebSocket 타입 정의
- `config/websocket.ts`: URL 및 재연결 설정

🎯 목적:
- Single Responsibility Principle 준수
- SSR 환경 완벽 지원
- 코드 복잡도 감소
- 재사용성 향상

📌 Related:
- Issue [#273]
- @ai-bot-review (https://github.com/jjaayy2222/flownote-mvp/pull/281#pullrequestreview-3643716686) - SSR Safety, Modularity, Simplification
