+++ web_ui/.env.example
@@ -6,3 +6,7 @@ NEXT_PUBLIC_API_BASE=http://localhost:8000
 # 개발 환경: ws://localhost:8000/ws
 # 프로덕션 환경: wss://your-domain.com/ws
 NEXT_PUBLIC_WS_URL=ws://localhost:8000/ws
+
+# WebSocket SSR 폴백 URL (서버 사이드 렌더링 환경용)
+# 프로덕션 SSR: wss://your-domain.com/ws
+NEXT_PUBLIC_WS_FALLBACK_URL=ws://localhost:8000/ws
